Split utility.js into smaller libraries
